fix: Time Buttons Should be Scroll not the Whole Right Side (#11412)
