+++ packages/features/bookings/index.ts
@@ -6,3 +6,4 @@ export {
   EventTitle,
 } from "./components/event-meta";
 export { AvailableTimes, AvailableTimesSkeleton } from "./components/AvailableTimes";
+export { AvailableTimesHeader } from "./components/AvailableTimesHeader";
